Added a sample BDD testing. - 1 Feature: Creating a new Repository - 2 Scenarios: Successful and Unsuccessful creation of a new Repository - Feature can be found in Feature directory.
Added a couple of new PageObjectModels.

Added a new Test Case file: CreateRepositoryTests

Added a new JSON data file that holds the New Repo name and description

Added a random string generator method to the BasePage

A few adjustments to other pages to comply with the new BDD testing technique.